more

+++ merge-dev-to-master.sh
@@ -1,19 +1,23 @@
 #!/bin/bash
-rm -f include/build_timestamp.h
-git checkout include/build_timestamp.h
+curbranch=`git branch | grep \* | cut -d ' ' -f 2`
 
-git checkout develop
-git pull
-git push
+if [[ x"$curbranch" == x"master" || x"$curbranch" == x"develop" ]]; then
+    rm -f include/build_timestamp.h
+    git checkout include/build_timestamp.h
 
-rm -f include/build_timestamp.h
-git checkout include/build_timestamp.h
+    git checkout develop
+    git pull
+    git push
 
-git checkout master
-git pull
-git push
+    rm -f include/build_timestamp.h
+    git checkout include/build_timestamp.h
 
-rm -f include/build_timestamp.h
-git checkout include/build_timestamp.h
+    git checkout master
+    git pull
+    git push
 
-git checkout master && git merge develop && git checkout develop && git merge master && git fetch --all && git push --all && make -j
+    rm -f include/build_timestamp.h
+    git checkout include/build_timestamp.h
+
+    git checkout master && git merge develop && git checkout develop && git merge master && git fetch --all && git push --all && make -j
+fi
